### 2022-02-10 #4152 @whedon
[PRE REVIEW]: dNami: a framework for solving systems of balance laws using explicit numerical schemes on structured meshes
**Submitting author:** @etouber (<a href="http://orcid.org/0000-0002-7553-0351">Emile Touber</a>)
**Repository:** <a href="https://github.com/dNamiLab/dNami" target ="_blank">https://github.com/dNamiLab/dNami</a>
**Version:** v1.0.0
