### 2022-02-11 87dfb2c
Update paper.bib
fixed the invalid or missing DOIs
