+++ JOSS/paper.bib
@@ -6,6 +6,7 @@
   number={1},
   pages={194--214},
   year={2004},
+  doi={10.1016/j.jcp.2003.09.003},
   publisher={Elsevier}
 }
 
@@ -135,6 +136,7 @@ archivePrefix = {arXiv},
   number={4}, 
   pages={296--305}, 
   year={2009}, 
+  doi={10.1504/IJCSE.2009.029165}, 
   publisher={Inderscience Publishers} 
 }
 
@@ -157,7 +159,7 @@ address = {Boston},
 pages = {377-396},
 year = {2015},
 isbn = {978-0-12-802118-7},
-doi = {https://doi.org/10.1016/B978-0-12-802118-7.00023-6},
+doi = {10.1016/B978-0-12-802118-7.00023-6},
 url = {https://www.sciencedirect.com/science/article/pii/B9780128021187000236},
 author = {Cedric Andreolli and Philippe Thierry and Leonardo Borges and Gregg Skinner and Chuck Yount},
 keywords = {Stencil, 3D finite differences, 3D application characterization, Arithmetic intensity, autotuning, finite difference, hpc, optimization, performance modeling, roofline model, Roofline model, LINPACK, Stream, wave equation, Xeon, Xeon Phi},
