### 2022-02-11 015b384
Update paper.bib
missing numba DOI
