+++ JOSS/paper.bib
@@ -38,7 +38,8 @@
   title={Numba: A llvm-based python jit compiler}, 
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley}, 
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC}, 
-  pages={1--6}, 
+  pages={1--6},
+  doi={10.1145/2833157.2833162}, 
   year={2015} 
 }
 
