### 2022-02-12 193322b
Update doc (correct 3D NS eqns)
