### 2022-02-12 f9d1695
Update doc (correct 2D NS eqns)
