### 2022-02-15 ed04677
Modified the restart function
