+++ src/src_py/dnami_io.py
@@ -353,22 +353,24 @@ def read_restart(tree,fname='restart.bin'):
 
         # read shell (if necessary)     
         fnameshell = {} 
-        if  len(bcs['allbc']) != 0:     
+        if  len(bcs['allbc']) != 0:
+                if ioproc:
+                        print('Reading boundary shell data from the restart file.')     
+                dirBC = []
+                if 'i1' in bcs['bcun']:
+                    dirBC = dirBC + ['i1','imax']
+                if 'j1' in bcs['bcun']:
+                    dirBC = dirBC + ['j1','jmax']
+                if 'k1' in bcs['bcun']:
+                    dirBC = dirBC + ['k1','kmax']
 
-                dirBC = ['i1','imax']
-                if ndim == 2:
-                        dirBC = dirBC + ['j1','jmax']
-                if ndim == 3:   
-                        dirBC = dirBC + ['j1','jmax','k1','kmax']
+                for direction in dirBC:
+                        fnameshell[direction] = 'restartshell_'+direction
 
-                for dir in dirBC:
-                        fnameshell[dir] = 'restartshell_'+dir
-
-        if fnameshell != {}     :
-
+        if fnameshell != {}:
                 position = {'i':[dmpi.ibeg,dmpi.iend],
-                                'j':[dmpi.jbeg,dmpi.jend],
-                                'k':[dmpi.kbeg,dmpi.kend]}
+                            'j':[dmpi.jbeg,dmpi.jend],
+                            'k':[dmpi.kbeg,dmpi.kend]}
 
                 nglb    = {'i':nxgb,'j':nygb,'k':nzgb}
 
@@ -376,8 +378,7 @@ def read_restart(tree,fname='restart.bin'):
                 header  = {}
                 subarray= {}
 
-                for dir in ['i1','imax','j1','jmax','k1','kmax']:
-
+                for idx,dire in enumerate(dirBC):
                         wposition = {'i':dmpi.ibeg,
                                      'j':dmpi.jbeg,
                                      'k':dmpi.kbeg}
@@ -398,7 +399,7 @@ def read_restart(tree,fname='restart.bin'):
 
 
                         for d in extpos:
-                                if d != dir[0]:
+                                if d != dire[0]:
                                         if position[d][0] == 1:
                                                 index[d][0]= 0                                  
                                                 sizeloc[d] = sizeloc[d] + hlo
@@ -408,64 +409,64 @@ def read_restart(tree,fname='restart.bin'):
                                                 index[d][1]= index[d][1] + hlo
                                                 sizeloc[d] = sizeloc[d]  + hlo
 
+                        sizeglb[dire[0]] = hlo
 
-                        # header
-                        headsize = 7; head = np.empty(headsize,dtype=wp)
-                        head[:]  = [headsize,n,t,nxgb,nygb,nzgb,nvar] # implicit type conversion        
-                
+                        head = np.empty(headsize,dtype=wp)
                         if iMpi:
-                                header[dir] = MPIWP.Create_contiguous(headsize)
-                                header[dir].Commit()    
-                                fh[dir] = MPI.File.Open(dmpi.comm_torus,fnameshell[dir],MPI.MODE_WRONLY|MPI.MODE_CREATE)
-                                fh[dir].Set_view(0,MPIWP,header[dir])
-                                if ioproc: fh[dir].Read_at(0,head)
-                                header[dir].Free()      
-                                fh[dir].Close()
-                                        
+                           dat     =     np.empty((sizeloc['i'],sizeloc['j'],sizeloc['k'],nvar),dtype=wp)
+                           header[dire] = MPIWP.Create_contiguous(headsize)
+                           header[dire].Commit()    
[…]
