### 2022-02-15 f280a29
added non-periodic bc test cases and fixes to the dnami_io read_restart [NRY]
