+++ exm/3d_io_bc/rhs.py
@@ -1,9 +1,9 @@
 # =============================================================================
-# 1D fake equations 
+# 3D fake equations 
 # =============================================================================
 
 # number of dimensions
-dim = 1 
+dim = 3 
 
 # coefficients ////////////////////////////////////////////////////////////////
 coefficients = {
@@ -43,4 +43,9 @@ locname_bc  = {'a' : 'bc_a',
 
 F = {'a' : ' 0.0_wp  ',
      'b' : ' 0.0_wp  '}
+
+# Boundary conditions /////////////////////////////////////////////////////////
+
+qbc = {'a' : ' 0.0_wp  ',
+       'b' : ' 0.0_wp  '}
 
