+++ exm/3d_io_periodic/rhs.py
@@ -1,5 +1,5 @@
 # =============================================================================
-# 3D navier stokes equations  
+# 3D fake equations 
 # =============================================================================
 
 # number of dimensions
