+++ exm/1d_io_bc/rhs.py
@@ -44,3 +44,8 @@ locname_bc  = {'a' : 'bc_a',
 F = {'a' : ' 0.0_wp  ',
      'b' : ' 0.0_wp  '}
 
+# Boundary conditions /////////////////////////////////////////////////////////
+
+qbc = {'a' : ' 0.0_wp  ',
+       'b' : ' 0.0_wp  '}
+

[…]
