### 2022-02-17 fc2e4f0
Added the OMP_NUM_THREADS=1 env. variable
