### 2022-02-20 7f294a0
Update dnami_io.py and fix a BUG in genBC
