### 2022-02-21 7acf7fc
fixed typos in the dnami_unpack bc function and full range of swap in swap function
