### 2022-02-21 #4152 @kthyng
Hi @mancellin and @jamiejquinn! I am looking for two reviewers for this JOSS submission. I would hope for reviews in 4–6 weeks, though typically starting earlier than that window because the review process tends to be iterative between the author(s) and reviewers. You can read more about the review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
Any interest?
