### 2022-02-22 #4186 @whedon
[REVIEW]: dNami: a framework for solving systems of balance laws using explicit numerical schemes on structured meshes
**Submitting author:** @etouber (<a href="http://orcid.org/0000-0002-7553-0351">Emile Touber</a>)
**Repository:** https://github.com/dNamiLab/dNami
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
