### 2022-02-22 #4186 @kthyng
Hi everyone! @JamieJQuinn, @mancellin, here is where the review will take place. As specific comments/issues come up, please open issues in the [software repo](https://github.com/dNamiLab/dNami) itself and link back to this review issue so we can track related issues/comments. You can track your overall review here with the checklist and ping the author(s) and me as needed, and otherwise let us know when you're finished.
