### 2022-02-22 #4186 @kthyng
@etouber Forgot in the pre-review issue: please make sure to add a section to your paper called "Statement of Need" — we like to be very upfront about the purpose of the software in the paper.
