### 2022-02-24 dNamiLab/dNami#18 @mancellin
Conserved variables?
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4186), I've started reading the tutorial and documentation of the code.

I am a bit confused about the implementation of the Euler equations in the example.
The files `rhs.py` and `genRhs.py` (which might have been more clearly named `equations.py` and `discretize.py`, but that's not the point) do not seem to mention the conserved variables `rho`, `rho u` and `rho e_t`, but only the base variables `rho`, `u` and `e_t`. I suspected it had something to do with the (undocumented?) `consvar`. After searching for it in the source code, it seems to add some lines of Fortran code with multiplications or divisions by `rho`.

I don't mind that there is a special case for variables of this form, since they are common for your use case in fluid dynamics, but it should be better documented.
For instance, users should be warned if it only works with a variable named `rho`.
By the way, does this make the scheme conservative or just add a transformation from one set of variables to the other?

Maybe a simpler test case without this kind of fine details would be a better fit for the quick-start guide?
