+++ JOSS/paper.md
@@ -43,7 +43,7 @@ in a flexible and efficient manner, where $\textbf{q} \in \mathbb{R}^n$ is a vec
 
 The ability of \texttt{dNami} to clearly separate the problem statement from its numerical implementation (often a major time sink in research laboratories) is rooted in the flexibility of the Python language so as to let the user define her/his own system of balance laws in the most natural way (i.e.\ using a human-readable syntax), which is then interpreted in Fortran to build a computationally-efficient library of \autoref{eq:gov_eq} which is callable from Python. Users can then easily interact with their own system of balance laws, including at runtime, thereby making it possible to integrate solutions to \autoref{eq:gov_eq} with other tools and libraries (e.g.\ optimisation and stability tools) to fully explore the properties of the system, seamlessly from small to large-scale calculations.
 
-# State of the field
+# Statement of Need  
 
 Code developments aimed at producing numerical solvers for \autoref{eq:gov_eq} typically follow two different strategies which usually involve two different communities. 
 
