+++ doc/usage/genRhs.rst
@@ -156,7 +156,9 @@ This ends the list of compulsory steps when creating a ``genRhs.py``.
         
 *Conservative formulation*
 
-dNami offers the user the possibility of advancing the equations in time using a conservative formulation. In the example equations given above, the 'solved' variables are ``rho``, ``u``, ``v`` and ``et`` but the quantities advanced in time by the Runge--Kutta scheme are ``rho``, ``rho*u``, ``rho*v`` and ``rho*et``, referred to as the conserved variables. In practice, the switch to conserved variables is usually done to leverage conservation properties of some finite-difference formulations and having access to the 'solved' variables is useful for setting initial conditions, boundary conditions and outputting information at run time.  This requires a transformation between the solved variables and the conserved variables before and after the Runge--Kutta steps. Currently, given a list of solved variables and a variable with the protected name ``rho`` e.g.    
+dNami offers the user the possibility of advancing the equations in time using a conservative formulation. In the example equations given above, the 'solved' variables are ``rho``, ``u``, ``v`` and ``et`` but the quantities advanced in time by the Runge--Kutta scheme are ``rho``, ``rho*u``, ``rho*v`` and ``rho*et``, referred to as the conserved variables. In practice, the switch to conserved variables is usually done to leverage conservation properties of some finite-difference formulations and having access to the 'solved' variables is useful for setting initial conditions, boundary conditions and outputting information at run time.  This requires a transformation between the solved variables and the conserved variables before and after the Runge--Kutta steps. Having access to the solved variables is crucial for efficient computation of the right-hand side of the time-advancement equations as it is more readily expressed as a function of the solved variables rather than the conserved variables. Substitutions of the form ``(rho*vari)/rho`` would be necessary and severely detrimental to efficiency if the switch was not performed. 
+
+Currently, given a list of solved variables and a variable with the protected name ``rho`` e.g.    
 
 ``varsolved = [rho, var1, var2, var3, var4, var5, ..., varN]``
 
