### 2022-03-02 #4186 @kthyng
@mancellin Ok sounds like we should try the new approach. Can you run this command at the top of a comment?

`@editorialbot generate my checklist`

This should generate your own checklist in line here in the comments that you will have access to. Assuming that works, I will then delete the list at the top of this issue, and then @JamieJQuinn would probably need to do the same thing. This is how the checklists will work moving forward to improve permissions issues, but we are right at the switchover time.
