### 2022-03-02 dNamiLab/dNami#19 @mancellin
Some suggestions for a more approachable code
Below are some comments I collected when going through the `genRhs.py` and `genKer.py` files. Most of them are very subjective, so take them as suggestions and not as requirements for publication of the [JOSS paper](https://github.com/openjournals/joss-reviews/issues/4186).

* To make the code more approachable by new users and developers, I suggest following the standard typographic convention of writing names of:
  * modules, functions and variables in snake_case,
  * classes in CamelCase,
  * global constants in CAPITALS.

* There are two different things called respectively `rhsinfo` and `rhs_info`. Maybe you could give them distinct names. For instance `Parameters` and `export_parameters`.

* In `rhs_info` in `genKer.py`, you import from `genRhs.py` some variables that were actually defined in `rhs.py`. Why not directly read `rhs.py`?
  More generally, you call in `genRhs.py` functions defined in `genKer.py` that import variables from `genRhs.py`. In my opinion, you are risking making your library an unmaintainable entangled mess, which could easily be avoided by passing variables (such as `incPath`) as function arguments.

* In many functions in `genKer.py`, `rhs` is an optional argument (with default value `None`). But the functions cannot succeed if `rhs` is not provided, so why not making it a mandatory argument and have more explicit error messages?

* I don't really understand the purpose of `Save_eqns`. Are the `append_Rhs` functions supposed to modify their first argument?
