### 2022-03-03 b1c77b2
Update documentation (genRhs.rst)
