+++ doc/usage/genRhs.rst
@@ -156,7 +156,14 @@ This ends the list of compulsory steps when creating a ``genRhs.py``.
         
 *Conservative formulation*
 
-dNami offers the user the possibility of advancing the equations in time using a conservative formulation. In the example equations given above, the 'solved' variables are ``rho``, ``u``, ``v`` and ``et`` but the quantities advanced in time by the Runge--Kutta scheme are ``rho``, ``rho*u``, ``rho*v`` and ``rho*et``, referred to as the conserved variables. In practice, the switch to conserved variables is usually done to leverage conservation properties of some finite-difference formulations and having access to the 'solved' variables is useful for setting initial conditions, boundary conditions and outputting information at run time.  This requires a transformation between the solved variables and the conserved variables before and after the Runge--Kutta steps. Having access to the solved variables is crucial for efficient computation of the right-hand side of the time-advancement equations as it is more readily expressed as a function of the solved variables rather than the conserved variables. Substitutions of the form ``(rho*vari)/rho`` would be necessary and severely detrimental to efficiency if the switch was not performed. 
+Conservation laws in physic can be written in multiple mathematically equivalent forms, yet numerical methods or implementation considerations may dictate specific choices. A classical formulation in continuum mechanics is the so-called divergence form:
+
+.. math::
+    \dfrac{\partial \vec{ \mathrm{q}}}{\partial t} + \overrightarrow{\mathrm{div}} \left( \overline{\overline{\mathrm{F}}} \right) = \overrightarrow{\mathrm{S}}
+
+, where the :math:`\vec{\mathrm{q}}` vector takes the form of :math:`\left[ \rho, \rho \mathrm{var1}, \rho \mathrm{var2},\cdots\right]^{\intercal}` and the flux tensor :math:`\overline{\overline{\mathrm{F}}}` is made explicit. Conservation of mass may then be ensured exactly with any finite-difference discretisation of the divergence operator, providing that appropriate numerical fluxes are defined based on :math:`\overline{\overline{\mathrm{F}}}` :cite:`vinokur1989`. The divergence form is therefore often referred as conservative formulation in continuum mechanics. 
+
+dNami offers the user the possibility of advancing the equations in time using a conservative formulation while minimising data transfer. In the Euler equations given above, the 'solved' variables, which are actually stored in memory, are ``rho``, ``u``, ``v`` and ``et`` but the quantities advanced in time by the Runge--Kutta scheme are ``rho``, ``rho*u``, ``rho*v`` and ``rho*et``, referred to as the conserved variables. In practice, having access to the 'solved' variables is useful for setting initial conditions, boundary conditions and outputting information at run time. This requires a transformation between the solved variables and the conserved variables before and after the Runge--Kutta steps. Having access to the solved variables is crucial for efficient computation of the right-hand side of the time-advancement equations as it is more readily expressed as a function of the solved variables rather than the conserved variables. Substitutions of the form ``(rho*vari)/rho`` would be necessary and severely detrimental to efficiency if the switch was not performed. 
 
 Currently, given a list of solved variables and a variable with the protected name ``rho`` e.g.    
 
