### 2022-03-04 12d0fbf
removing save_eqns step and doc tweaks
