+++ exm/1d_euler_nonreflective/genRhs.py
@@ -26,14 +26,12 @@ def main():
     genrk3update(len(varsolved),rhs=rhs)
 
 # Generate RHS:
-    Save_eqns = {'divF':divF.copy()}
     append_Rhs(divF, 5,4, rhsname, vnamesrc_divF, update=False,rhs=rhs,stored=True)                           
-
 # Generate Filters (if required):      
     genFilter(11,10, len(varsolved),rhs=rhs)
 
 # Progressive stencil/order adjustement from domain to boundary 
-    genBC(Save_eqns['divF']  ,3,2, rhsname , vnamesrc_divF, update=False,rhs=rhs)
+    genBC(divF,3,2, rhsname , vnamesrc_divF, update=False,rhs=rhs)
 
 # Boundary conditions on d(q)/dt 
     #i1
