### 2022-03-04 d76acf7
updated the post I/O functions, restored the shell indexing, modified the io tests to rest the shell files
