+++ exm/1d_io_bc/compute.py
@@ -103,7 +103,7 @@ dMpi.swap(q,hlo,dtree)
 if dMpi.ioproc:
     for shell in glob.glob('restartshell*'):
         rname = shell.split('_')
-        rname = rname[0] + '_' + rname[-1]
+        rname = rname[0] + '_' + rname[-2]
         shutil.copyfile(shell, rname)
 
 # -- Read it back in  
