### 2022-03-04 6d07021
update genRhs.rst
