+++ exm/2d_vortex_exit/genRhs.py
@@ -1,7 +1,7 @@
 import sys
 import re
 import numpy as np
-from genKer import rhsinfo, genrk3, genrk3update, genFilter, genBC, append_Rhs, genbcsrc
+from genKer import genrk3, genrk3update, genFilter, genBC, append_Rhs
 import os 
 
 wp = 'float64'
@@ -54,7 +54,7 @@ def main():
     genBC(src_phybc_wave_imaxjmax,5,4,rhsname , locname_bc, setbc=[True,{'cornerchar':{'imaxjmax':['rhs']}}]  , update=False,rhs=rhs)
 
     # Extract RHS info:
-    rhsinfo(rhs)
+    rhs.export()
 
 if __name__ == '__main__':
     main()
