+++ src/generate/genKer.py
@@ -22,7 +22,16 @@ except:
  
 class rhs_info:
 	def __init__(self):
-		from genRhs	import dim, wp
+		try: 
+			from genRhs	import dim
+		except:
+			exception('dim not define. Try fixing the number of spatial dimension (dim = 1, dim = 2 or dim = 3)',message='error')
+
+		try: 
+			from genRhs	import wp
+		except:
+			exception('wp not defined. Try fixing the working precision',message='error')
+
 
 		try:
 			from genRhs import consvar
@@ -54,6 +63,15 @@ class rhs_info:
 		except:
 			exception('varname not defined',message='error')
 
+		try:
+			from genRhs import hlo_glob
+		except:
+			exception('The number of hlo cells, hlo_glob, needs to be defined in genRhs. It is not automatically computed yet.',message='error')								   
+		try:
+			from genRhs import incPATH
+		except:
+			exception('incPATH not defined in genRhs',message='error')	
+		
 		self.wp     	  = wp     	 
 		self.dim     	  = dim     	 
 		self.stencil 	  = 1 	 
@@ -64,8 +82,45 @@ class rhs_info:
 		self.varstored    = varstored  
 		self.varsolved    = varsolved
 		self.varbc        = varbc
+		self.varloc	      = varloc
 		self.hlo_rhs      = 1
+		self.hlo_glob     = hlo_glob
 		self.bc_info      = [{},{}]
+		self.incPATH      = incPATH
+
+	def export(self):
+		"""
+			 rhsinfo is a user-level function that needs to be called at the end of genRhs.py in order to write necessary rhs information to be used at runtime.
+			 """
+		genBC_calls(self)
+
+		wp     	     = self.wp     	     
+		dim     	 	 = self.dim     	     
+		stencil 	 	 = self.stencil 	     
+		order        = self.order         
+		coefficients = self.coefficients  
+		varname      = self.varname   
+		varsolved    = self.varsolved
+		varstored    = self.varstored    
+		hlo_rhs      = self.hlo_rhs  
+		bc_info      = self.bc_info     
+		varbc        = self.varbc
+
+
+		instpath = os.environ['INSTALLPATH']
+		rhsinf = open(instpath+'/pymod/rhsinfo.py','w')
+
+		rhsinf.write('wp = \''+str(wp)+'\'\n')
+		rhsinf.write('dim = '+str(dim)+'\n')
+		rhsinf.write('stencil = '+str(stencil)+'\n')
+		rhsinf.write('order = '+str(order)+'\n')
+		rhsinf.write('coefficients = '+str(coefficients)+'\n')
+		rhsinf.write('varname = '+str(varname)+'\n')
+		rhsinf.write('varsolved = '+str(varsolved)+'\n')	
+		rhsinf.write('varstored = '+str(varstored)+'\n')
+		rhsinf.write('hlo_rhs = '+str(hlo_rhs)+'\n')
+		rhsinf.write('bc_info = '+str(bc_info)+'\n')
+		rhsinf.write('varbc = '+str(varbc)+'\n')	
 
 
 
@@ -287,7 +342,7 @@ Variables  = lambda input: [s.strip() for s in opsplit(input)]
 Operations = lambda input: re.findall(oplist,input)
 
 
-def dNamiVar(var,rangei,rangej,rangek):
+def dNamiVar(var,rangei,rangej,rangek,rhs):
 	"""
[…]
