+++ exm/3d_tgv/genRhs.py
@@ -1,7 +1,7 @@
 import sys
 import re
 import numpy as np
-from genKer import rhsinfo, genrk3, genrk3update, genFilter, genBC, append_Rhs, genbcsrc
+from genKer import genrk3, genrk3update, genFilter, genBC, append_Rhs
 import os 
 
 wp = 'float64'
@@ -33,7 +33,7 @@ def main():
       genFilter(13,8, len(varsolved),rhs=rhs)
 
 # Extract RHS info:
-      rhsinfo(rhs)
+      rhs.export()
 
 if __name__ == '__main__':
     main()
