+++ exm/1d_euler_nonreflective/genRhs.py
@@ -1,7 +1,7 @@
 import sys
 import re
 import numpy as np
-from genKer import rhsinfo, genrk3, genrk3update, genFilter, genBC, append_Rhs, genbcsrc
+from genKer import genrk3, genrk3update, genFilter, genBC, append_Rhs
 import os 
 
 # Set working precision
@@ -40,7 +40,7 @@ def main():
     genBC(src_phybc_wave_imax,3,2,rhsname ,vnamesrc_divF, setbc=[True,{'char':{'imax':['rhs']}}]  , update=False,rhs=rhs)
 
     # Extract RHS info:
-    rhsinfo(rhs)
+    rhs.export()
 
 if __name__ == '__main__':
     main()
