### 2022-03-05 b0bc08a
updated rhs.export docstring
