### 2022-03-12 b24c82c
fixed shell varible wrapping in env_dNami and fixed the error message when number of processors in compute differ from requested number
