### 2022-03-14 7880a4d
changes in the examples folder renaming rhs.py to equations.py
