+++ exm/1d_entropywave/genRhs.py
@@ -6,7 +6,7 @@ import os
 
 wp = 'float64'
   
-from rhs import *
+from equations import *
 
 # Set install path (needed by genKer)
 instpath = os.environ['INSTALLPATH']
