### 2022-03-14 9e3b9aa
update the documentation to reflection the rhs.py>equations.py changes
