+++ doc/usage/syntax.rst
@@ -7,7 +7,7 @@ user can easily translate differential equations into finite difference code.
 Floats 
 ------
 
-When adding a float to the pseudo-code in an ``rhs.py`` file, it should be followed by the suffix ``_wp`` which makes sure the floating point constant has the correct working precision chosen in the ``genRhs.py``. For example:
+When adding a float to the pseudo-code in an ``equations.py`` file, it should be followed by the suffix ``_wp`` which makes sure the floating point constant has the correct working precision chosen in the ``genRhs.py``. For example:
 
 .. code-block:: python 
 
