+++ doc/usage/testcases.rst
@@ -309,7 +309,7 @@ The incompressible pressure solution is projected onto an isochore in thermodyna
 
       Animation of the x-direction velocity field at :math:`z=z_{max}/2`
 
-The more common low-Mach validation case with :math:`Ma=0.1` is run. The complexity of the case arises as the flow transitions to turbulence. A lack of spatial resolution quickly leads to a degradation of the solution due to dissipation and dispersion. A more detailed discussion can be found here :cite:`debonis2013solutions`. A commonly used measure involves tracking the enstrophy (i.e. the domain integral of the squared vorticity) over time. This is strongly affected by numerical methods (e.g. the amount of filtering or the numerical dissipation introduced by the discretisation). With dNami pseudo-code, different formulations of the governing equations can easily be implemented. The ``rhs.py`` file for the 3D TGV case contains two versions of the equations. :numref:`3d_tgv_formulation` shows a comparison between conservative and skew-symmetric formulations for various grid sizes and a comparison to a spectral method based reference :cite:`wang2013high` of the enstrophy  over reduced time. All three finite-difference based computations presented in the graph use an 11 point, 10\ :sup:`th` order scheme.  
+The more common low-Mach validation case with :math:`Ma=0.1` is run. The complexity of the case arises as the flow transitions to turbulence. A lack of spatial resolution quickly leads to a degradation of the solution due to dissipation and dispersion. A more detailed discussion can be found here :cite:`debonis2013solutions`. A commonly used measure involves tracking the enstrophy (i.e. the domain integral of the squared vorticity) over time. This is strongly affected by numerical methods (e.g. the amount of filtering or the numerical dissipation introduced by the discretisation). With dNami pseudo-code, different formulations of the governing equations can easily be implemented. The ``equations.py`` file for the 3D TGV case contains two versions of the equations. :numref:`3d_tgv_formulation` shows a comparison between conservative and skew-symmetric formulations for various grid sizes and a comparison to a spectral method based reference :cite:`wang2013high` of the enstrophy  over reduced time. All three finite-difference based computations presented in the graph use an 11 point, 10\ :sup:`th` order scheme.  
 
 .. _3d_tgv_formulation:
 .. figure:: img/3d_tgv_enstrophy.png

[…]
