### 2022-03-14 dNamiLab/dNami#18 @etouber
Hi @mancellin, thank you for raising these points. We have updated the documentation to include the missing information by adding a sub-section about how to use the conservative formulation within the dNami framework in the 'Writing your own genRhs.py' section.

We believe the Euler equations to be a good first example because they allow us to showcase how to solve a multi-component equation with a few 'intermediate' variables (internal energy, pressure) as well as boundary conditions. The conservative formulation aspect should be useful to communities outside of fluid mechanics, i.e. more generally to continuum mechanics fields.
