### 2022-03-14 dNamiLab/dNami#19 @etouber
Hi @mancellin, thank you for the suggestions. Regarding your first point, this task requires some significant work to unify the code and risks breaking some functionalities in the short-run. We have a few additional test cases in mind we would like to prepare with the current version that would then be used to validate a reformatting down the line.

Regarding your other points:

    - rhs.py is now called equations.py in the documentation and the examples. Note that this file is not mandatory to the dNami framework. The only requirement is that dictionaries of equations are seen at the genRhs.py level (part of the framework).

    - we have added an ``export`` method to the rhs class to remove the confusion between rhsinfo and rhs_info

    - all imports from genRhs.py are now performed once when instantiating the rhs_info class. Explicit error messages are added when rhs_info is not instantiated before a call to append_Rhs or genBC to reflect anything the user might have missed. This should streamline the 'genRhs.py' file. For example, the user no longer needs to carry out the 'Save_eqns' step with these modifications, as you pointed out.

Thank you for helping us improve dNami!
