### 2022-03-16 dNamiLab/dNami#19 @mancellin
Thank you for taking my suggestions into account! The code seems already clearer to me.

If you don't mind, I'd like to suggest you to go one step further: remove the import statements in `rhs_info` and pass explicitly the parameters, as in
```python
rhs = rhs_info(varname=varname, varsolved=varsolved,
               hlo_glob=6, dim=1, wp='float64', incPATH='...')
```
It is a bit more verbose, but I believe it is worth it:
* it is more explicit and users know where each parameter is required (e.g. it is currently hard to guess that `wp` should be defined *before* calling `rhs_info()`).
* it should work even if `genRhs.py` is not named `genRhs`, and even if it is not in the same directory as `genKer.py` (as long as `genKer.py` is in the PYTHONPATH),
* you could then be a few small steps away from removing the requirement of manually copying files in `generate/` in dNami source tree, making dNami less cumbersome for users and for testing.
