### 2022-03-16 #4186 @mancellin
dNami a very interesting, ambitious and meaningful work.

I was able to consult the documentation, install the code and run some test cases, without any major issue.
I don't have the hardware to really benchmark the HPC performance of the code, but the claims seems reasonable given my small scale tests.
I submitted a few minor issues, that are being addressed.

Overall, the code and the paper meet all the requirements, and **I can recommend the paper for publication**.

Nonetheless, I have some concerns about the long-term maintainability of the code. Code-writing code is hard to debug, and the design of dNami is not always simple and straightforward (e.g. up to 15 levels of indentations in genKer.py). I would suggest that the authors take some time to make sure of the correctness, readability and maintainability of the code. For instance, by adding a more extensive test suite including unit tests. Refactoring dNami into a pip-installable library that write and compile code in a directory that is not the sources of the library itself, could also make it more convenient both for users and for tests. I don't feel like making this kind of large refactoring a condition for the publication of this already good work, and leave this only as a suggestion to the authors.
