### 2022-03-18 83b9f90
Update rhs_info class
