+++ exm/2d_wavy_mesh/genRhs.py
@@ -21,7 +21,13 @@ hlo_glob = 5
 def main():
       
     from genKer import rhs_info    
-    rhs = rhs_info()
+    # rhs = rhs_info(dim,wp,hlo_glob,incPATH,varsolved,varname, 
+    #                consvar=consvar,varstored=varstored,varloc=varloc,varbc=varbc,
+    #                coefficients=coefficients) 
+    
+    rhs = rhs_info(dim,wp,hlo_glob,incPATH,varsolved,varname,  
+                   consvar=consvar,varstored=varstored,varloc=varloc,
+                   coefficients=coefficients)   
 
 # Generate LHS:
     genrk3(len(varsolved)      ,rhs=rhs) 
