+++ exm/3d_io_bc/genRhs.py
@@ -18,7 +18,13 @@ hlo_glob = 5
 def main():
 
       from genKer import rhs_info
-      rhs = rhs_info()
+    # rhs = rhs_info(dim,wp,hlo_glob,incPATH,varsolved,varname, 
+    #                consvar=consvar,varstored=varstored,varloc=varloc,varbc=varbc,
+    #                coefficients=coefficients) 
+    
+      rhs = rhs_info(dim,wp,hlo_glob,incPATH,varsolved,varname,  
+                     consvar=consvar,varloc=varloc,
+                     coefficients=coefficients)
 
 # Generate LHS:
       genrk3(len(varsolved)      ,rhs=rhs)
