### 2022-03-19 d391147
updated documentation to account for rhs class changes
