### 2022-03-19 7dbc8db
small updates to readme and paper
