+++ JOSS/paper.md
@@ -55,7 +55,7 @@ A relatively recent trend aimed at maintaining HPC capabilities whilst providing
 
 # Features 
 
-At the core of \texttt{dNami} is the translation of symbolic expressions written in high-level Python language to discretise equations in low-level Fortran language. \texttt{dNami} employs explicit schemes to discretise differential operators. For the temporal derivative of \autoref{eq:gov_eq}, a low-storage 3$^\mathrm{rd}$ order Runge--Kutta (RK) scheme is used (other explicit schemes may easily be implemented). Spatial derivatives are discretised using finite differences of arbitrary orders provided by the user. A choice between standard and optimised schemes [such as those in @bogey2004family] is available. Both the governing equation in the form of "$\partial\textbf{q}/\partial t = f(\textbf{q})$" and the boundary conditions are specified symbolically. \texttt{dNami} automatically deals with stencil-size and order reduction close to boundaries using the user-specified symbolic equations, which removes the need for time-consuming and often problem-specific code development.
+At the core of \texttt{dNami} is the translation of symbolic expressions written in high-level Python language to discretised equations in low-level Fortran language. \texttt{dNami} employs explicit schemes to discretise differential operators. For the temporal derivative of \autoref{eq:gov_eq}, a low-storage 3$^\mathrm{rd}$ order Runge--Kutta (RK) scheme is used (other explicit schemes may easily be implemented). Spatial derivatives are discretised using finite differences of arbitrary orders provided by the user. A choice between standard and optimised schemes [such as those in @bogey2004family] is available. Both the governing equation in the form of "$\partial\textbf{q}/\partial t = f(\textbf{q})$" and the boundary conditions are specified symbolically. \texttt{dNami} automatically deals with stencil-size and order reduction close to boundaries using the user-specified symbolic equations, which removes the need for time-consuming and often problem-specific code development.
 
 The source-to-source translation is performed by a set of Python functions using regular expressions (see `genKer.py`). The produced discretised version of \autoref{eq:gov_eq} is then inserted into appropriate do-loops included in Fortran template files by pre-processing techniques. This simple yet effective strategy makes it possible for the HPC-layer to be tailored at the template-file level independently of \autoref{eq:gov_eq}. Finally, the resulting Fortran source code is compiled as a shared library and optimised through the auto-optimisation process of modern Fortran compilers. A Python module is then created with a high-level interface of the shared library functions using F2PY [@2009f2py]. It is important to stress that researchers can still follow a traditional development workflow, in a pure Fortran environment, either at the shared library level or inside the high-level interface. 
 
