### 2022-03-24 e64c348
Removed equation tags from the doc because the pdf doc didn't compile correctly
