### 2022-03-24 9d3eafc
The dnami syntax table caused the latex doc build to fail, added a conditional statement to distinguish between the html and pdf doc
