### 2022-04-04 3382fb8
1d standing shock case added to exm and start of documentation
