### 2022-04-11 3ef5e8d
changes to check if hlo_glob is reasonable or not
