### 2022-06-27 73bb57c
Update .readthedocs.yaml
Changed the Python version to avoid a build error for the doc.
