+++ .readthedocs.yaml
@@ -5,7 +5,7 @@ version: 2
 build:
   os: "ubuntu-20.04"
   tools:
-    python: "3.9"
+    python: "3.9.0"
 
 # Build from the docs/ directory with Sphinx
 sphinx:
