### 2022-06-27 c2f13b5
Update .readthedocs.yaml
Changed the version back to 3.9.
