### 2022-06-27 1ab2cf3
Update requirements.txt
Added a specific version of the library to resolve the readthedocs build error.
