+++ doc/requirements.txt
@@ -1,3 +1,4 @@
+funcparserlib == 1.0.0a0
 sphinx<4
 sphinx-panels
 sphinx_copybutton
