### 2022-06-27 caeb996
Update .readthedocs.yaml
Changes the version to resolve the build error.
