### 2022-06-27 7dfb9ee
Update requirements.txt
Changed the version of setuptools.
