### 2022-06-27 6a60792
Update requirements.txt
Changed the version to avoid build errors for the doc.
