+++ doc/requirements.txt
@@ -1,4 +1,4 @@
-setuptools==58.0
+setuptools<58.0
 sphinx<4
 sphinx-panels
 sphinx_copybutton
