### 2022-06-27 f636e31
Update requirements.txt
Changed the version of setuptools.
