+++ doc/requirements.txt
@@ -1,4 +1,4 @@
-setuptools==57.5.0
+pip install setuptools==58
 sphinx<4
 sphinx-panels
 sphinx_copybutton
