### 2022-06-27 7da03d8
Update requirements.txt
Removed setuptools.
