### 2022-06-27 3b448d7
Update .readthedocs.yaml
Changes the Python version.
