### 2022-06-27 32a26b3
Update .readthedocs.yaml
Changed the Python version back to 3.9.
